requirements.txt
biopython==1.65
boltons==0.6.4
Flask==0.10.1
PyYAML==3.11
requests==2.7.0
